Qonos API rate control POST jobs requests
We need to keep the API from falling over when too many
POST workers/jobs requests are incoming.  To do this we
leverage the rate limit'ing feature of nova-api to keep
from too many requests being processed back through the database.

Rather than returning and error, we are electing to return "no job".
The worker is designed to retry periodically which should spread
them out.

RM9089